+++ src/Actions/SetMaterial.lua
@@ -1,6 +1,0 @@
-return function(material)
-  return {
-    type = 'SetMaterial',
-    material = material
-  }
-end