+++ src/Actions/SetPanel.lua
@@ -1,6 +1,0 @@
-return function(panel)
-  return {
-    type = 'SetPanel',
-    panel = panel
-  }
-end