+++ src/Actions/init.lua
@@ -1,6 +1,0 @@
-local Actions = {
-  SetMaterial = require(script.SetMaterial),
-  SetPanel = require(script.SetPanel)
-}
-
-return Actions