+++ src/Components/App.lua
@@ -6,36 +6,21 @@ local MaterialPanel = require(Modules.MaterialPanel)
 local SettingsPanel = require(Modules.SettingsPanel)
 local ThemedTextLabel = require(Modules.ThemedTextLabel)
 local Localization = require(Modules.Parent.Util.Localization)
+local Store = require(Modules.Store)
 
 local App = Roact.PureComponent:extend('App')
 
-function App:init()
-  local _store = self.props.store
-
-  self.state = {
-    store = _store:getState()
-  }
-
-  _store.changed:connect(function()
-    self:setState({
-      store = _store:getState()
-    })
-  end)
-end
-
 function App:render()
   local props = self.props
-  local state = self.state
-  local store = state.store
+  local store = self.state.Panel
 
   local body
 
-  if store.Panel == 'Materials' then
+  if store == 'Materials' then
     body = Roact.createElement(MaterialPanel, {
       Size = props.IsOutdated and UDim2.new(1,-10,1,-53) or UDim2.new(1,-10,1,-35),
-      MaterialSelected = store.Material
     })
-  elseif store.Panel == 'Settings' then
+  elseif store == 'Settings' then
     body = Roact.createElement(SettingsPanel, {
       Size = props.IsOutdated and UDim2.new(1,-10,1,-53) or UDim2.new(1,-10,1,-35),
       plugin = props.plugin
@@ -53,9 +38,7 @@ function App:render()
         Padding = UDim.new(0, 5),
         SortOrder = 'LayoutOrder'
       }),
-      Navbar = Roact.createElement(Navbar, {
-        ActiveFrame = store.Panel
-      }),
+      Navbar = Roact.createElement(Navbar),
       Body = body,
 
       update = props.IsOutdated and Roact.createElement(ThemedTextLabel, {
@@ -68,4 +51,4 @@ function App:render()
   end)
 end
 
-return App
+return Store:Roact(App, { 'Panel' })