+++ src/Components/MaterialPanel/Item.lua
@@ -1,7 +1,8 @@
 local Modules = script.Parent.Parent.Parent
 local Roact = require(Modules.Libs.Roact)
-local RoactRodux = require(Modules.Libs.RoactRodux)
-local Actions = require(Modules.Actions)
+
+local Store = require(script.Parent.Parent.Store)
+
 local Constants = require(Modules.Util.Constants)
 local Localization = require(Modules.Util.Localization)
 local ThemedTextLabel = require(script.Parent.Parent.ThemedTextLabel)
@@ -12,7 +13,7 @@ local function getTextSize(Text)
   return UDim2.new(0, tb.X + 3, 0, tb.Y + 3)
 end
 
-local function MaterialItem(props)
+local function MaterialItem()
   local elements = {}
 
   for i=1, #Constants.MATERIALS do
@@ -21,7 +22,7 @@ local function MaterialItem(props)
       BackgroundTransparency = 1,
       Image = item.img,
       [Roact.Event.MouseButton1Click] = function()
-        props.SetMaterial(item.enum)
+        Store:Set('Material', item.enum)
       end,
       [Roact.Event.MouseEnter] = function(rbx)
         local tooltip = rbx:WaitForChild('Tooltip')
@@ -41,7 +42,7 @@ local function MaterialItem(props)
         Image = Constants.MATERIAL_SELECTED_IMAGE,
         Size = UDim2.new(0.5,0,0.5,0),
         Position = UDim2.new(0.5,0,0,0),
-        Visible = props.MaterialSelected == item.enum
+        Visible = Store:Get('Material') == item.enum
       }),
       Tooltip = Roact.createElement(ThemedTextLabel, {
         BackgroundTransparency = 0.3,
@@ -58,16 +59,4 @@ local function MaterialItem(props)
   return Roact.createFragment(elements)
 end
 
-local function mapStateToProps()
-  return {}
-end
-
-local function mapDispatchToProps(dispatch)
-  return {
-    SetMaterial = function(mat)
-      dispatch(Actions.SetMaterial(mat))
-    end
-  }
-end
-
-return RoactRodux.connect(mapStateToProps,mapDispatchToProps)(MaterialItem)
+return MaterialItem