+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Navbar is now centered with a fixed width to the buttons.
 * Adjusted the first time size and the minimum size for plugin widget.
 * Increased the size of material images from 45px to 50px
+* Replaced Rodux with BasicState
 
 ## [1.4.0] - July 11, 2020
 ### Added